+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 - results layer with customizable result views
 
 ### Changed
+- update and add new icons
+- make data in legend dynamically and add legend schema and example
 - extend popup schema with key values component
 - load satellite layer beneath symbol layers #43
 - switch top left logo for final version